requirements.txt
fastapi
uvicorn
python-dotenv
beanie